+++ lib/xcodeproj/project.rb
@@ -539,32 +539,6 @@ module Xcodeproj
       main_group.new_group(name, path, source_tree)
     end
 
-    # Adds a file reference for a system framework to the project if needed and
-    # adds it to the {PBXFrameworksBuildPhase} of the given target.
-    #
-    # @example Adding QuartzCore
-    #
-    #         framework = project.add_system_framework('QuartzCore', :ios)
-    #
-    #         target = project.targets.first
-    #         build_phase = target.frameworks_build_phases.first
-    #         build_phase.files << framework.buildFiles.new
-    #
-    # @param  [String] name
-    #         The name of a framework.
-    #
-    # @param  [PBXNativeTarget] target
-    #         The target to which to add the framework.
-    #
-    # @note   This method adds a reference to the highest know SDK for the
-    #         given platform.
-    #
-    # @return [PBXFileReference] The file reference of the framework.
-    #
-    def add_system_framework(name, target)
-      ProjectHelper.add_system_framework(self, name, target)
-    end
-
     # Creates a new target and adds it to the project.
     #
     # The target is configured for the given platform and its file reference it